% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cggm.R
\name{cggm}
\alias{cggm}
\title{Clusterpath Estimator of the Gaussian Graphical Model}
\usage{
cggm(
  S,
  W_cpath,
  lambda_cpath,
  W_lasso = NULL,
  lambda_lasso = 0,
  eps_lasso = 0.005,
  gss_tol = 0.005,
  conv_tol = 1e-07,
  fusion_threshold = NULL,
  tau = 0.001,
  max_iter = 5000,
  expand = FALSE,
  max_difference = 0.01,
  verbose = 0
)
}
\arguments{
\item{S}{The sample covariance matrix of the data.}

\item{W_cpath}{The weight matrix used in the clusterpath penalty.}

\item{lambda_cpath}{A numeric vector of tuning parameters for regularization.
Should be a sequence of monotonically increasing values.}

\item{W_lasso}{The weight matrix used in the lasso penalty. Defaults to
\code{NULL}, which is interpreted as all weights being zero (no
penalization).}

\item{lambda_lasso}{The penalty parameter used for the lasso penalty.
Defaults to \code{0} (no penalization).}

\item{eps_lasso}{Parameter that governs the quadratic approximation of the
lasso penalty. Within the interval \code{c(-eps_lasso, eps_lasso)} the
absolute value function is approximated by a quadratic function. Defaults to
\code{5e-3}.}

\item{gss_tol}{The tolerance value used in the Golden Section Search (GSS)
algorithm. Defaults to \code{1e-6}.}

\item{conv_tol}{The tolerance used to determine convergence. Defaults to
\code{1e-7}.}

\item{fusion_threshold}{The threshold for fusing two clusters. If
\code{NULL}, defaults to \code{tau} times the median distance between the
rows of \code{solve(S)}.}

\item{tau}{The parameter used to determine the fusion threshold. Defaults to
\code{1e-3}.}

\item{max_iter}{The maximum number of iterations allowed for the optimization
algorithm. Defaults to \code{5000}.}

\item{expand}{Determines whether the vector \code{lambda} should be expanded
with additional values in order to find a sequence of solutions that (a)
terminates in the minimum number of clusters and (b) has consecutive
solutions for Theta that are not too different from each other. The degree
of difference between consecutive solutions that is allowed is determined by
\code{max_difference}. Defaults to \code{FALSE}.}

\item{max_difference}{The maximum allowed difference between consecutive
solutions of Theta if \code{expand = TRUE}. The difference is computed as
\code{norm(Theta[i-1]-Theta[i], "F") / norm(Theta[i-1], "F")}. Defaults to
\code{0.01}.}

\item{verbose}{Determines the amount of information printed during the
optimization. Slows down the algorithm significantly. Defaults to \code{0}.}
}
\value{
An object of class \code{"CGGM"} with the following components:
\item{A,R}{Lists of matrices. Each pair of matrices with the same index
parametrize the estimated precision matrix for the corresponding value
of the aggregation parameter \code{lambda_cpath}. It is not recommended to
use these directly, instead use the accessor function
\code{\link{get_Theta}()} to extract the estimated precision matrix for a
given index of the aggregation parameter.}
\item{clusters}{An integer matrix in which each row contains the cluster
assignment of each variable for the corresponding value of the aggregation
parameter \code{lambda_cpath}. Use the accessor function
\code{\link{get_clusters}()} to extract the cluster assignment for a given
index of the aggregation parameter.}
\item{lambdas}{A vector with the values for the regularization parameter
lambda that the  CGGM loss function has been  minimized for.}
\item{Theta}{List of matrices. Contains the solution to the minimization
procedure for each value for lambda. It is not recommended to use these
directly, instead use the accessor function \code{\link{get_Theta}()} to
extract the estimated precision matrix for a given index of the aggregation
parameter.}
\item{losses}{A vector with the values of the minimized CGGM loss function
for each value for lambda.}
\item{cluster_counts}{An integer vector containing the number of clusters
obtained for each value for lambda.}
\item{loss_progression}{A list of vectors. Contains, for each value for
lambda, the value of the loss function for each iteration of the
minimization procedure. Only part of the output if \code{expand = FALSE}.}
\item{loss_timings}{A list of vectors. Contains, for each value for lambda,
the cumulative elapsed computation time for each iteration of the
minimization procedure.}
\item{fusion_threshold}{The threshold value used to determine whether two
clusters should be clustered.}
\item{cluster_solution_index}{An integer vector containing the index of the
value for lambda for which a certain number of clusters was attained. For
example, \code{cluster_solution_index[2]} yields the index of the smallest
value for lambda for which a solution with two clusters was found. Contains
-1 if there is no value for lambda with that number of clusters.}
\item{n}{The number of values for lambda for which the CGGM loss function
was minimized.}
\item{inputs}{A list of the inputs of the function, used internally and in
\code{\link{cggm_refit}()}. It consists of eight components:
\itemize{
\item{S} (the sample covariance matrix)
\item{W_cpath} (the weight matrix for the clusterpath penalty)
\item{gss_tol} (the tolerance for the GSS algorithm)
\item{conv_tol} (the convergence tolerance)
\item{max_iter} (the maximum number of iterations)
\item{lambda_lasso} (the penalty parameter for the lasso penalty)
\item{eps_lasso} (parameter used for the quadratic approximation of the
lasso penalty)
\item{W_lasso} (the weight matrix for the lasso penalty)
}}
}
\description{
Compute the clusterpath estimator of the Gaussian Graphical Model (CGGM) for
fixed values of the tuning parameters to obtain a sparse estimate with
variable clustering of the precision matrix or the covariance matrix.
}
\note{
The function interface and output structure are still experimental and
may change in the next version.
}
\references{
D.J.W. Touw, A. Alfons, P.J.F. Groenen and I. Wilms (2025)
\emph{Clusterpath Gaussian Graphical Modeling}. arXiv:2407.00644.
doi:10.48550/arXiv.2407.00644.
}
\seealso{
\code{\link{clusterpath_weights}()}, \code{\link{lasso_weights}()},
\code{\link{cggm_refit}()}, \code{\link{cggm_cv}()}
}
\author{
Daniel J.W. Touw
}
