% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cggm_cv.R
\name{cggm_cv}
\alias{cggm_cv}
\title{Cross Validation for the Clusterpath Gaussian Graphical Model}
\usage{
cggm_cv(
  X,
  tune_grid,
  kfold = 5,
  folds = NULL,
  connected = TRUE,
  fit = TRUE,
  refit = TRUE,
  lasso_unit_weights = FALSE,
  estimate_Sigma = FALSE,
  verbose = 0,
  n_jobs = 1,
  ...
)
}
\arguments{
\item{X}{The \code{n} times \code{p} matrix holding the data, with \code{n}
observations and \code{p} variables.}

\item{tune_grid}{A data frame with values of the tuning parameters. Each row
is a combination of parameters that is evaluated. The columns have the names
of the tuning parameters and should include \code{k} and \code{phi}. The
regularization parameter \code{lambda} is optional. If there is no column
named \code{lambda}, an appropriate range is selected for each combination of
\code{k} and \code{phi}.}

\item{kfold}{The number of folds. Defaults to 5.}

\item{folds}{Optional argument to manually set the folds for the cross
validation procedure. If this is not \code{NULL}, it overrides the
\code{kfold} argument. Defaults to \code{NULL}.}

\item{connected}{Logical, indicating whether connectedness of the weight
matrix should be ensured. Defaults to \code{TRUE}. See
\code{\link{cggm_weights}}.}

\item{fit}{Logical, indicating whether the cross-validation procedure
should consider the result from \code{\link{cggm}}, before refitting is
applied. Defaults to \code{TRUE}. At least one of \code{fit} and \code{refit}
should be \code{TRUE}.}

\item{refit}{Logical, indicating whether the cross-validation procedure
should also consider the refitted result from \code{\link{cggm}}. See also
\code{\link{cggm_refit}}. Defaults to \code{TRUE}. At least one of \code{fit}
and \code{refit} should be \code{TRUE}.}

\item{lasso_unit_weights}{Logical, indicating whether the weights in the
sparsity penalty should be all one or decreasing in the magnitude of the
corresponding element of the inverse of the sample covariance matrix.
Defaults to \code{FALSE}.}

\item{estimate_Sigma}{Logical, indicating whether CGGM should be used to
estimate the covariance matrix based on the sample precision matrix. Defaults
to \code{FALSE}.}

\item{verbose}{Determines the amount of information printed during the
cross validation. Defaults to \code{0}.}

\item{n_jobs}{Number of parallel jobs used for cross validation. If 0 or
smaller, uses the maximum available number of physical cores. Defaults to
\code{1} (sequential).}

\item{...}{Additional arguments meant for \code{\link{cggm}} and
\code{\link{cggm_refit}}.}
}
\value{
A list containing the estimated parameters of the CGGM model.
}
\description{
This function performs k-fold cross validation to tune the weight matrix
parameters phi and k (for k nearest neighbors, knn) and the regularization
parameters lambda_cpath and lambda_lasso. The scoring metric is the negative
log-likelihood (lower is better).
}
\examples{
# Example usage:

}
\seealso{
\code{\link{clusterpath_weights}}, \code{\link{lasso_weights}},
\code{\link{cggm}}
}
