% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cggm_expand.R
\name{cggm_expand}
\alias{cggm_expand}
\title{Estimate Clusterpath Gaussian Graphical Model}
\usage{
cggm_expand(cggm_output, lambdas, verbose = 0)
}
\arguments{
\item{cggm_output}{Output of the cggmNew function.}

\item{lambdas}{Additional lambdas for which the loss should be minimized.
Values for which there already is a solution are discarded as well as those
that are smaller than the smallest value: \code{min(cggm_output$lambdas)}.}

\item{verbose}{Determines the amount of information printed during the
optimization. Defaults to \code{0}.}
}
\value{
A list containing the estimated parameters of the CGGM model.
}
\description{
This function estimates expands a previous solutionpath by minimizing the
loss function for additional values for lambda. It uses warm starts extracted
from a previous result to reduce computational burden.
}
\examples{
# Example usage:

}
