% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cggm_refit.R
\name{cggm_refit}
\alias{cggm_refit}
\title{Refit the Gaussian Graphical Model for a Given Aggregation and Sparsity Structure}
\usage{
cggm_refit(cggm_output, verbose = 0)
}
\arguments{
\item{cggm_output}{An object of class \code{"CGGM"} as returned by
\code{\link{cggm}()}.}

\item{verbose}{Determines the amount of information printed during the
optimization. Defaults to \code{0}.}
}
\value{
An object of class \code{"CGGM_refit"} with the following components:
\item{A,R}{Lists of matrices. Each pair of matrices with the same index
parametrize the estimated precision matrix after the refitting step given
the aggregation structure found with the corresponding value of the
aggregation parameter \code{lambda_cpath} (and sparsity structure found with
the value of the sparsity parameter \code{lambda_lasso}). It is not
recommended to use these directly, instead use the accessor function
\code{\link{get_Theta}()} to extract the estimated precision matrix for a
given index of the aggregation parameter.}
\item{clusters}{An integer matrix in which each row contains the cluster
assignment of each variable for the corresponding value of the aggregation
parameter \code{lambda_cpath}. Use the accessor function
\code{\link{get_clusters}()} to extract the cluster assignment for a given
index of the aggregation parameter.}
\item{lambdas}{A vector with the values for the regularization parameter
lambda that the  CGGM loss function has been  minimized for.}
\item{Theta}{List of matrices. Contains the solution to the minimization
procedure for each value for lambda. It is not recommended to use these
directly, instead use the accessor function \code{\link{get_Theta}()} to
extract the estimated precision matrix for a given index of the aggregation
parameter.}
\item{cluster_counts}{An integer vector containing the number of clusters
obtained for each value for lambda.}
\item{cluster_solution_index}{An integer vector containing the index of the
value for lambda for which a certain number of clusters was attained. For
example, \code{cluster_solution_index[2]} yields the index of the smallest
value for lambda for which a solution with two clusters was found. Contains
-1 if there is no value for lambda with that number of clusters.}
\item{n}{The number of values for lambda for which the CGGM loss function
was minimized.}
}
\description{
Estimate the parameters of a clustered and sparse precision matrix or
covariance matrix based on a restricted negative log-likelihood loss
function. The restrictions are given by the provided aggregation and
sparsity structure. This function is different from \code{\link{cggm}()},
as there are no aggregation and sparsity penalties on the precision or
covariance matrix.
}
\note{
The function interface and output structure are still experimental and
may change in the next version.
}
\references{
D.J.W. Touw, A. Alfons, P.J.F. Groenen and I. Wilms (2025)
\emph{Clusterpath Gaussian Graphical Modeling}. arXiv:2407.00644.
doi:10.48550/arXiv.2407.00644.
}
\seealso{
\code{\link{cggm}()}, \code{\link{cggm_cv}()}
}
\author{
Daniel J.W. Touw
}
