% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cggm_weights.R
\name{cggm_weights}
\alias{cggm_weights}
\title{Compute weight matrix for the Clusterpath Gaussian Graphical Model}
\usage{
cggm_weights(S, phi, k, connected = TRUE)
}
\arguments{
\item{S}{The sample covariance matrix of the data.}

\item{phi}{Tuning parameter of the weights.}

\item{k}{The number of nearest neighbors that should be used to set weights
to a nonzero value. If \code{0 < k < ncol(S)}, the dense weight matrix will
be made sparse, otherwise the dense matrix is returned.}

\item{connected}{Boolean that indicates whether a connected weight matrix
should be enforced. Defaults to \code{TRUE}.}
}
\value{
A weight matrix.
}
\description{
This function computes the (possibly sparse) weight matrix that is used in
CGGM. Weights are computed using exp{-phi * norm(Theta, i, j)^2}, where the
squared norm of two rows of Theta are scaled by the average squared norm of
all the rows of Theta. Additionally, groups of variables that would not be
connected via nonzero weights due to the sparsity of the weight matrix can
still be connected by applying a minimum spanning tree algorithm.
}
\examples{
# Example usage:

}
