% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cggmCV.R
\name{cggmCV}
\alias{cggmCV}
\title{Cross Validation for the Clusterpath Gaussian Graphical Model}
\usage{
cggmCV(
  X,
  lambdas,
  phi,
  k,
  kfold = 5,
  folds = NULL,
  cov_method = "pearson",
  gss_tol = 1e-04,
  conv_tol = 1e-09,
  fusion_type = "proximity",
  fusion_threshold = NULL,
  max_iter = 5000,
  use_Newton = TRUE,
  scoring_method = "NLL",
  check_k = TRUE
)
}
\arguments{
\item{X}{The sample covariance matrix of the data.}

\item{lambdas}{A numeric vector of tuning parameters for regularization.
Make sure the values are monotonically increasing.}

\item{phi}{A numeric vector of tuning parameters for the weights.}

\item{k}{A vector of integers for the number of nearest neighbors that
should be used to set weights to a nonzero value.}

\item{kfold}{The number of folds. Defaults to 5.}

\item{folds}{Optional argument to manually set the folds for the cross
validation procedure. If this is not \code{NULL}, it overrides the
\code{kfold} argument. Defaults to \code{NULL}.}

\item{cov_method}{Character string indicating which correlation coefficient
(or covariance) is to be computed. One of \code{"pearson"}, \code{"kendall"},
or \code{"spearman"}: can be abbreviated. Defaults to \code{"pearson"}.}

\item{gss_tol}{The tolerance value used in the Golden Section Search (GSS)
algorithm. Defaults to \code{1e-4}.}

\item{conv_tol}{The tolerance used to determine convergence. Defaults to
\code{1e-9}.}

\item{fusion_type}{The type of fusion to be used for determining fusion
between two objects. Possible values: "proximity" (based on closeness) or
one of {"a0", "a1", "a2"} (based on analytical evaluation). Defaults to
"proximity".}

\item{fusion_threshold}{If \code{fusion_type = "proximity"}, it is the
threshold for fusing two clusters. For the analytical fusions, it is the
threshold below which the analytical check is executed. If NULL, defaults
to \code{sqrt(nrow(S)) * 1e-4} times the median distance between the rows of
\code{solve(S)} for proximity based fusions and to \code{sqrt(nrow(S))} times
the median distance for analytical fusions.}

\item{max_iter}{The maximum number of iterations allowed for the optimization
algorithm. Defaults to \code{5000}.}

\item{use_Newton}{Logical, indicating whether to use Newton's method in the
optimization algorithm. Defaults to \code{TRUE}.}

\item{scoring_method}{Method to use for the cross validation scores. The
choices are negative log likelihood (\code{NLL}), \code{AIC}, and \code{BIC}.
Defaults to \code{NLL}. There is also \code{Test}, which is for testing other
options.}

\item{check_k}{Logical, indicating whether to check whether some elements of
\code{k} yield the same dense weight matrix. If so, values leading to
duplicates of the weight matrix are eliminated from the cross validation
grid. Defaults to \code{TRUE}.}
}
\value{
A list containing the estimated parameters of the CGGM model.
}
\description{
This function performs k-fold cross validation to tune the weight matrix
parameters phi and k (for k nearest neighbors, knn) and the regularization
parameter lambda.
}
\examples{
# Example usage:

}
