% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cggmRefit.R
\name{cggmRefit}
\alias{cggmRefit}
\title{Estimate Clusterpath Gaussian Graphical Model}
\usage{
cggmRefit(cggm_output, verbose = 0, ...)
}
\arguments{
\item{cggm_output}{Output of \code{cggmNew()}.}

\item{verbose}{Determines the amount of information printed during the
optimization. Defaults to \code{0}.}
}
\value{
A list containing the estimated parameters of the CGGM model.
}
\description{
This function estimates the parameters of a CGGM model based on the input
data and parameters.
}
\examples{
# Example usage:

}
