% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cggmWeights.R
\name{cggmWeights}
\alias{cggmWeights}
\title{Compute weight matrix for the Clusterpath Gaussian Graphical Model}
\usage{
cggmWeights(S, phi, k)
}
\arguments{
\item{S}{The sample covariance matrix of the data.}

\item{phi}{Tuning parameter of the weights.}

\item{k}{The number of nearest neighbors that should be used to set weights
to a nonzero value. If \code{0 < k < ncol(S)}, the dense weight matrix will
be made sparse, otherwise the dense matrix is returned.}
}
\value{
A weight matrix.
}
\description{
This function computes the (possibly sparse) weight matrix that is used in
CGGM. Weights are computed using exp{-phi * norm(Theta, i, j)^2}, where the
squared norm of two rows of Theta are scaled by the average squared norm of
all the rows of Theta.
}
\examples{
# Example usage:

}
