% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterGGM-package.R
\docType{package}
\name{clusterGGM-package}
\alias{clusterGGM}
\alias{clusterGGM-package}
\title{\packageTitle{clusterGGM}}
\description{
\packageDescription{clusterGGM}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{clusterGGM}
\packageIndices{clusterGGM}
}
\references{
D.J.W. Touw, A. Alfons, P.J.F. Groenen and I. Wilms (2025)
\emph{Clusterpath Gaussian Graphical Modeling}. arXiv:2407.00644.
doi:10.48550/arXiv.2407.00644.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/aalfons/clusterGGM}
  \item Report bugs at \url{https://github.com/aalfons/clusterGGM/issues}
}

}
\author{
\packageAuthor{clusterGGM}

Maintainer: \packageMaintainer{clusterGGM}
}
\keyword{package}
