% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterglasso_permutation.R
\name{clusterglasso_permutation}
\alias{clusterglasso_permutation}
\title{cluster-glasso estimation of the precision matrix with permutation matrix such that diagonal elements are taken into account}
\usage{
clusterglasso_permutation(
  X,
  W = NULL,
  pendiag = F,
  lambda1,
  adaptive = FALSE,
  power_adaptive = 1,
  W_sparsity = NULL,
  lambda2,
  knn_weights = F,
  knn_connect = F,
  phi = 1,
  knn = 3,
  eps_fusions = 0.001,
  rho = 10^-2,
  it_in = 100,
  it_out = 10,
  refitting = T,
  it_in_refit = 100,
  it_out_refit = 10
)
}
\arguments{
\item{X}{An (\eqn{n}x\eqn{p})-matrix of \eqn{p} variables and \eqn{n} observations}

\item{W}{An (\eqn{p}x\eqn{p})-matrix of weights}

\item{pendiag}{Logical indicator whether or not to penalize the diagonal in Omega. The default is \code{FALSE} (no penalization of the diagonal)}

\item{lambda1}{Sparsity tuning parameter.}

\item{adaptive}{Logical indicator whether an adaptive lasso type of sparsity penalty should be used or not. Default is FALSE}

\item{power_adaptive}{Power for the weights int he adaptive lasso type of sparsity penalty. Default is one.}

\item{W_sparsity}{An (\eqn{p}x\eqn{p})-matrix of weights for the sparsity penalty term. Only relevant if adaptive = TRUE}

\item{lambda2}{Aggregation tuning parameter.}

\item{knn_weights}{Boolean to indicate whether knn weights are used.}

\item{phi}{A scalar to tune the weights in the W for clustering}

\item{knn}{The number of nearest neighbors used for the W for clustering.}

\item{eps_fusions}{Threshold for fusing clusters. Default is 10^-3}

\item{rho}{Starting value for the LA-ADMM tuning parameter. Default is 10^2; will be locally adjusted via LA-ADMM}

\item{it_in}{Number of inner stages of the LA-ADMM algorithm. Default is 100}

\item{it_out}{Number of outer stages of the LA-ADMM algorithm. Default is 10}

\item{refitting}{Logical indicator whether refitting subject to sparsity and aggregation constraints is done or not. The default is \code{TRUE}.}

\item{it_in_refit}{Number of inner stages of the LA-ADMM algorithm for re-fitting (Only relevant if \code{refitting = TRUE}). Default is 100}

\item{it_out_refit}{Number of outer stages of the LA-ADMM algorithm for re-fitting (Only relevant if \code{refitting = TRUE}). Default is 10}
}
\value{
A list with the following components
\item{\code{omega_full}}{Estimated (\eqn{p}x\eqn{p}) precision matrix}
\item{\code{cluster}}{Numeric vector indicating the cluster groups for each of the \eqn{p} original variables}
\item{\code{sparsity}}{The (\eqn{p}x\eqn{p} matrix indicating the sparsity pattern in Omega (1=non-zero, 0=zero))}
\item{\code{fit}}{Fitted object from LA_ADMM_clusterglasso_export cpp function, for internal use now}
\item{\code{refit}}{Fitted object from refit_LA_ADMM_export cpp function, for internal use now}
}
\description{
This function computes the cluster-lasso estimator (thereby also accounting for the diagonal elements when clustering) for fixed tuning parameters lambda1 and lambda2
}
