% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_folds.R
\name{cv_folds}
\alias{cv_folds}
\title{Create Cross-Validation Folds}
\usage{
cv_folds(n, K = 5L)
}
\arguments{
\item{n}{an integer giving the number of observations to be split.}

\item{K}{an integer giving the number of blocks into which the observations
should be split (the default is five).}
}
\value{
A list of indices giving the blocks of observations to be folded
into training and test data during cross-validation.
}
\description{
Obtain indices for splitting observations into \eqn{K} blocks to be to be
folded into training and test data during \eqn{K}-fold cross-validation.
}
\examples{
cv_folds(20, K = 5)

}
\seealso{
\code{\link{cggm_cv}()}
}
\author{
Andreas Alfons
}
