% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCovariance.R
\name{generateCovariance}
\alias{generateCovariance}
\title{Generate Covariance Matrix}
\usage{
generateCovariance(n_vars, n_clusters, n_draws = 100 * n_vars, shuffle = FALSE)
}
\arguments{
\item{n_vars}{The number of variables in the covariance matrix.}

\item{n_clusters}{The number of clusters in the covariance matrix.}

\item{n_draws}{The number of observations that should be drawn from a
multivariate normal distribution with the true covariance matrix.}

\item{shuffle}{Logical indicating whether to shuffle the covariance matrix.
Default = FALSE.}
}
\value{
A list containing the true covariance matrix and the sample
covariance matrix based on the specified parameters.
}
\description{
This function generates a \code{n_vars} by \code{n_vars} covariance matrix in
which the rows/columns are genereated from \code{n_clusters} clusters.
}
