% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_clusters}
\alias{get_clusters}
\alias{get_clusters.CGGM}
\alias{get_clusters.CGGM_refit}
\alias{get_clusters.CGGM_CV}
\title{Extract the Cluster Assignment}
\usage{
get_clusters(object, ...)

\method{get_clusters}{CGGM}(object, index, ...)

\method{get_clusters}{CGGM_refit}(object, index, ...)

\method{get_clusters}{CGGM_CV}(object, which = NULL, ...)
}
\arguments{
\item{object}{an object from which to extract the cluster assignment.}

\item{\dots}{additional arguments are currently ignored.}

\item{index}{an integer specifying the step along the clusterpath for which
to extract the cluster assignment.}

\item{which}{a character string specifying for which solution to extract
the cluster assignment.  Possible values are \code{"refit"} for the solution
including the refitting step (see \code{\link{cggm_refit}()}), or
\code{"fit"} for the solution without without the refitting step (see
\code{\link{cggm}()}).  If \code{NULL} (the default), the solution with the
better cross-validation score is used.}
}
\value{
An integer vector giving the obtained cluster assignment for each variable.

For the \code{"CGGM_CV"} method (see \code{\link{cggm_cv}()}), the returned
cluster assignment corresponds to the optimal values of the tuning
parameters.
}
\description{
Extract a cluster assignment obtained via the clusterpath estimator of the
Gaussian graphical model (CGGM).
}
\references{
D.J.W. Touw, A. Alfons, P.J.F. Groenen and I. Wilms (2025)
\emph{Clusterpath Gaussian Graphical Modeling}. arXiv:2407.00644.
doi:10.48550/arXiv.2407.00644.
}
\seealso{
\code{\link{cggm}()}, \code{\link{cggm_refit}()}, \code{\link{cggm_cv}()}

\code{\link{get_Theta}()}
}
\author{
Daniel J.W. Touw
}
