% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_weights.R
\name{lasso_weights}
\alias{lasso_weights}
\title{Compute weight matrix for the lasso penalty of the Clusterpath Gaussian
Graphical Model}
\usage{
lasso_weights(S, unit = FALSE)
}
\arguments{
\item{S}{The sample covariance matrix of the data.}

\item{unit}{Logical, indicating whether the weights should be all one or
based on the inverse of \code{S}.}
}
\value{
A weight matrix for the lasso penalty.
}
\description{
Compute weight matrix for the lasso penalty of the Clusterpath Gaussian
Graphical Model
}
\examples{
# Example usage:

}
