% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_clusters.R
\name{min_clusters}
\alias{min_clusters}
\title{Calculate the Minimum Number of Clusters}
\usage{
min_clusters(W)
}
\arguments{
\item{W}{The weight matrix for the clusterpath penalty.}
}
\value{
An integer giving the minimum number of clusters.
}
\description{
Compute the minimum number of clusters achievable by the clusterpath penalty
using the provided weight matrix.
}
\references{
D.J.W. Touw, A. Alfons, P.J.F. Groenen and I. Wilms (2025)
\emph{Clusterpath Gaussian Graphical Modeling}. arXiv:2407.00644.
doi:10.48550/arXiv.2407.00644.
}
\seealso{
\code{\link{clusterpath_weights}()}, \code{\link{cggm}()},
\code{\link{cggm_cv}()}
}
\author{
Daniel J.W. Touw
}
