% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterglasso.R
\name{preliminaries_for_DOC_subproblem}
\alias{preliminaries_for_DOC_subproblem}
\title{tag-lasso estimation of the precision matrix}
\usage{
preliminaries_for_DOC_subproblem(p)
}
\arguments{
\item{X}{An (\eqn{n}x\eqn{p})-matrix of \eqn{p} variables and \eqn{n} observations}

\item{pendiag}{Logical indicator whether or not to penalize the diagonal in Omega. The default is \code{TRUE} (penalization of the diagonal)}

\item{lambda1}{Sparsity tuning parameter.}

\item{lambda2}{Aggregation tuning parameter.}

\item{rho}{Starting value for the LA-ADMM tuning parameter. Default is 10^2; will be locally adjusted via LA-ADMM}

\item{it_in}{Number of inner stages of the LA-ADMM algorithm. Default is 100}

\item{it_out}{Number of outer stages of the LA-ADMM algorithm. Default is 10}

\item{it_in_refit}{Number of inner stages of the LA-ADMM algorithm for re-fitting. Default is 100}

\item{it_out_refit}{Number of outer stages of the LA-ADMM algorithm for re-fitting. Default is 10}
}
\value{
A list with the following components
\item{\code{omega_full}}{Estimated (\eqn{p}x\eqn{p}) precision matrix}
\item{\code{omega_aggregated}}{Estimated (\eqn{K}x\eqn{K}) precision matrix}
\item{\code{cluster}}{Numeric vector indicating the cluster groups for each of the \eqn{p} original variables}
\item{\code{M}}{The (\eqn{p}x\eqn{K} membership matrix)}
}
\description{
This function computes the tag-lasso estimator for fixed tuning parameters lambda1 and lambda2
}
