% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning_clusterglasso.R
\name{tuning_clusterglasso}
\alias{tuning_clusterglasso}
\title{tuning parameter selection for cluster-glasso with Bayesian Optimization}
\usage{
tuning_clusterglasso(
  X,
  lambda1_lb = 0,
  lambda1_ub = 0.2,
  lambda2_lb = 0,
  lambda2_ub = 0.6,
  bo_init = NULL,
  fold = 5,
  seed_cv,
  bo_iter = 10
)
}
\arguments{
\item{X}{An (\eqn{n}x\eqn{p})-matrix of \eqn{p} variables and \eqn{n} observations}

\item{lambda1_lb}{Lower bound in the domain of the sparsity tuning parameter}

\item{lambda1_ub}{Upper bound in the domain of the sparsity tuning parameter}

\item{lambda2_lb}{Lower bound in the domain of the aggregation tuning parameter}

\item{lambda2_ub}{Upper bound in the domain of the aggregation tuning parameter}

\item{bo_init}{Initialization matrix for Bayesian optimization. We advise to let the program determine the initialization matrix based on the provided lower and upper bounds of the tuning parameters}

\item{fold}{Folds to be used in cross-validation procedure. Default is 5-fold cross-validation}

\item{seed_cv}{Seed for creating folds in cross-validation procedure}

\item{bo_iter}{Number of iterations for Bayesian optimization.
#' @return A list with the following components
\item{\code{lambda1_opt}}{Selected sparsity tuning parameter}
\item{\code{lambda2_opt}}{Selected aggregation tuning parameter}}
}
\description{
This function selects the tuning parameter for the cluster-glasso with Bayesian Optimization
}
