% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cggm.R
\name{cggm}
\alias{cggm}
\title{Estimate Clusterpath Gaussian Graphical Model}
\usage{
cggm(
  S,
  W_cpath,
  lambda_cpath,
  W_lasso = NULL,
  lambda_lasso = 0,
  eps_lasso = 0.005,
  gss_tol = 0.005,
  conv_tol = 1e-07,
  fusion_threshold = NULL,
  tau = 0.001,
  max_iter = 5000,
  expand = FALSE,
  max_difference = 0.01,
  verbose = 0
)
}
\arguments{
\item{S}{The sample covariance matrix of the data.}

\item{W_cpath}{The weight matrix used in the clusterpath penalty.}

\item{lambda_cpath}{A numeric vector of tuning parameters for regularization.
Should be a sequence of monotonically increasing values.}

\item{W_lasso}{The weight matrix used in the lasso penalty. Defaults to
\code{NULL}, which is interpreted as all weights being zero (no
penalization).}

\item{lambda_lasso}{The penalty parameter used for the lasso penalty.
Defaults to \code{0} (no penalization).}

\item{eps_lasso}{Parameter that governs the quadratic approximation of the
lasso penalty. Within the interval \code{c(-eps_lasso, eps_lasso)} the
absolute value function is approximated by a quadratic function. Defaults to
\code{5e-3}.}

\item{gss_tol}{The tolerance value used in the Golden Section Search (GSS)
algorithm. Defaults to \code{1e-6}.}

\item{conv_tol}{The tolerance used to determine convergence. Defaults to
\code{1e-7}.}

\item{fusion_threshold}{The threshold for fusing two clusters. If
\code{NULL}, defaults to \code{tau} times the median distance between the
rows of \code{solve(S)}.}

\item{tau}{The parameter used to determine the fusion threshold. Defaults to
\code{1e-3}.}

\item{max_iter}{The maximum number of iterations allowed for the optimization
algorithm. Defaults to \code{5000}.}

\item{expand}{Determines whether the vector \code{lambda} should be expanded
with additional values in order to find a sequence of solutions that (a)
terminates in the minimum number of clusters and (b) has consecutive
solutions for Theta that are not too different from each other. The degree
of difference between consecutive solutions that is allowed is determined by
\code{max_difference}. Defaults to \code{FALSE}.}

\item{max_difference}{The maximum allowed difference between consecutive
solutions of Theta if \code{expand = TRUE}. The difference is computed as
\code{norm(Theta[i-1]-Theta[i], "F") / norm(Theta[i-1], "F")}. Defaults to
\code{0.01}.}

\item{verbose}{Determines the amount of information printed during the
optimization. Slows down the algorithm significantly. Defaults to \code{0}.}
}
\value{
A list containing the estimated parameters of the CGGM model.
}
\description{
This function estimates the parameters of a CGGM model based on the input
data and parameters.
}
\examples{
# Example usage:

}
\seealso{
\code{\link{clusterpath_weights}}, \code{\link{lasso_weights}}
}
