% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cggm.R
\name{cggm}
\alias{cggm}
\title{Estimate Clusterpath Gaussian Graphical Model}
\usage{
cggm(
  S,
  W,
  lambdas,
  gss_tol = 1e-04,
  conv_tol = 1e-09,
  fusion_type = "proximity",
  fusion_threshold = NULL,
  max_iter = 2000,
  store_all_res = FALSE,
  use_Newton = TRUE,
  profile = FALSE,
  verbose = 0
)
}
\arguments{
\item{S}{The sample covariance matrix of the data.}

\item{W}{The weight matrix used in the clusterpath penalty.}

\item{lambdas}{A numeric vector of tuning parameters for regularization.
Make sure the values are monotonically increasing.}

\item{gss_tol}{The tolerance value used in the Golden Section Search (GSS)
algorithm. Defaults to \code{1e-4}.}

\item{conv_tol}{The tolerance used to determine convergence. Defaults to
\code{1e-9}.}

\item{fusion_type}{The type of fusion to be used for determining fusion
between two objects. Possible values: "proximity" (based on closeness) or
one of {"a0", "a1", "a2"} (based on analytical evaluation). Defaults to
"proximity".}

\item{fusion_threshold}{If \code{fusion_type = "proximity"}, it is the
threshold for fusing two clusters. For the analytical fusions, it is the
threshold below which the analytical check is executed. If NULL, defaults
to \code{sqrt(nrow(S)) * 1e-4} times the median distance between the rows of
\code{solve(S)} for proximity based fusions and to \code{sqrt(nrow(S))} times
the median distance for analytical fusions.}

\item{max_iter}{The maximum number of iterations allowed for the optimization
algorithm. Defaults to \code{2000}.}

\item{store_all_res}{Logical, indicating whether to store the results for all
values for lambda. If false, only solutions for different numbers of
clusters are stored. Defaults to \code{FALSE}.}

\item{use_Newton}{Logical, indicating whether to use Newton's method in the
optimization algorithm. Defaults to \code{TRUE}.}

\item{profile}{Logical, indicates whether a profiling report should be
printed. Defaults to \code{FALSE}.}

\item{verbose}{Determines the amount of information printed during the
optimization. Slows down the algorithm significantly. Defaults to \code{0}.}
}
\value{
A list containing the estimated parameters of the CGGM model.
}
\description{
This function estimates the parameters of a CGGM model based on the input
data and parameters.
}
\examples{
# Example usage:

}
