% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cggm.R
\name{cggm}
\alias{cggm}
\title{Estimate Clusterpath Gaussian Graphical Model}
\usage{
cggm(
  S,
  W,
  lambdas,
  gss_tol = 1e-06,
  conv_tol = 1e-07,
  fusion_threshold = NULL,
  tau = 0.001,
  max_iter = 5000,
  store_all_res = FALSE,
  verbose = 0
)
}
\arguments{
\item{S}{The sample covariance matrix of the data.}

\item{W}{The weight matrix used in the clusterpath penalty.}

\item{lambdas}{A numeric vector of tuning parameters for regularization.
Make sure the values are monotonically increasing.}

\item{gss_tol}{The tolerance value used in the Golden Section Search (GSS)
algorithm. Defaults to \code{1e-6}.}

\item{conv_tol}{The tolerance used to determine convergence. Defaults to
\code{1e-7}.}

\item{fusion_threshold}{The threshold for fusing two clusters. If NULL,
defaults to \code{tau} times the median distance between the rows of
\code{solve(S)}.}

\item{tau}{The parameter used to determine the fusion threshold. Defaults to
\code{1e-3}.}

\item{max_iter}{The maximum number of iterations allowed for the optimization
algorithm. Defaults to \code{5000}.}

\item{store_all_res}{Logical, indicating whether to store the results for all
values for lambda. If false, only solutions for different numbers of
clusters are stored. Defaults to \code{FALSE}.}

\item{verbose}{Determines the amount of information printed during the
optimization. Slows down the algorithm significantly. Defaults to \code{0}.}
}
\value{
A list containing the estimated parameters of the CGGM model.
}
\description{
This function estimates the parameters of a CGGM model based on the input
data and parameters.
}
\examples{
# Example usage:

}
