% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cggm_cv.R
\name{cggm_cv}
\alias{cggm_cv}
\title{Cross Validation for the Clusterpath Gaussian Graphical Model}
\usage{
cggm_cv(
  X,
  tune_grid,
  kfold = 5,
  folds = NULL,
  connected = FALSE,
  scoring_method = "NLL",
  ...
)
}
\arguments{
\item{X}{The sample covariance matrix of the data.}

\item{tune_grid}{A data frame with values of the tuning parameters. Each row
is a combination of parameters that is evaluated. The columns have the names
of the tuning parameters and should include \code{k} and \code{phi}. The
regularization parameter \code{lambda} is optional. If there is no column
named \code{lambda}, an appropriate range is selected for each combination of
\code{k} and \code{phi} \code{lambda}.}

\item{kfold}{The number of folds. Defaults to 5.}

\item{folds}{Optional argument to manually set the folds for the cross
validation procedure. If this is not \code{NULL}, it overrides the
\code{kfold} argument. Defaults to \code{NULL}.}

\item{connected}{Logical, indicating whether connectedness of the weight
matrix should be ensured. Defaults to \code{FALSE}. See
\code{\link{cggm_weights}}.}

\item{scoring_method}{Method to use for the cross validation scores.
Currently, the only choice is \code{NLL} (negative log-likelihood).}

\item{...}{Additional arguments meant for \code{\link{cggm}}.}
}
\value{
A list containing the estimated parameters of the CGGM model.
}
\description{
This function performs k-fold cross validation to tune the weight matrix
parameters phi and k (for k nearest neighbors, knn) and the regularization
parameter lambda.
}
\examples{
# Example usage:

}
\seealso{
\code{\link{cggm_weights}}, \code{\link{cggm}}
}
