% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cggm_cv.R
\name{cggm_cv}
\alias{cggm_cv}
\title{Cross Validation for the Clusterpath Estimator of the Gaussian Graphical Model}
\usage{
cggm_cv(
  X,
  tune_grid,
  kfold = 5,
  folds = NULL,
  connected = TRUE,
  fit = TRUE,
  refit = TRUE,
  lasso_unit_weights = FALSE,
  estimate_Sigma = FALSE,
  verbose = 0,
  n_jobs = 1,
  ...
)
}
\arguments{
\item{X}{The \code{n} times \code{p} matrix holding the data, with \code{n}
observations and \code{p} variables.}

\item{tune_grid}{A data frame with values of the tuning parameters. Each row
is a combination of parameters that is evaluated. The columns have the names
of the tuning parameters and should include \code{k} and \code{phi}. The
regularization parameter \code{lambda} is optional. If there is no column
named \code{lambda}, an appropriate range is selected for each combination of
\code{k} and \code{phi}.}

\item{kfold}{The number of folds. Defaults to 5.}

\item{folds}{Optional argument to manually set the folds for the cross
validation procedure. If this is not \code{NULL}, it overrides the
\code{kfold} argument. Defaults to \code{NULL}.}

\item{connected}{Logical, indicating whether connectedness of the weight
matrix should be ensured. Defaults to \code{TRUE}. See
\code{\link{clusterpath_weights}()}.}

\item{fit}{Logical, indicating whether the cross-validation procedure
should consider the result from \code{\link{cggm}()}, before refitting is
applied. Defaults to \code{TRUE}. At least one of \code{fit} and \code{refit}
should be \code{TRUE}.}

\item{refit}{Logical, indicating whether the cross-validation procedure
should also consider the refitted result from \code{\link{cggm}()}. See also
\code{\link{cggm_refit}()}. Defaults to \code{TRUE}. At least one of
\code{fit} and \code{refit} should be \code{TRUE}.}

\item{lasso_unit_weights}{Logical, indicating whether the weights in the
sparsity penalty should be all one or decreasing in the magnitude of the
corresponding element of the inverse of the sample covariance matrix.
Defaults to \code{FALSE}.}

\item{estimate_Sigma}{Logical, indicating whether CGGM should be used to
estimate the covariance matrix based on the sample precision matrix. Defaults
to \code{FALSE}.}

\item{verbose}{Determines the amount of information printed during the
cross validation. Defaults to \code{0}.}

\item{n_jobs}{Number of parallel jobs used for cross validation. If 0 or
smaller, uses the maximum available number of physical cores. Defaults to
\code{1} (sequential).}

\item{...}{Additional arguments to be passed down to \code{\link{cggm}()}
and \code{\link{cggm_refit}()}.}
}
\value{
A list containing the estimated parameters of the CGGM model.
}
\description{
Perform cross validation to tune the weight matrix parameters \code{phi}
and \code{k} (for k-nearest-neighbors) as well as the aggregation parameter
\code{lambda_cpath} and the sparsity parameter \code{lambda_lasso} of the
clusterpath estimator of the Gaussian Graphical Model (CGGM) in order to
obtain a sparse estimate with variable clustering of the precision matrix or
the covariance matrix. The scoring metric is the negative log-likelihood
(lower is better).
}
\references{
D.J.W. Touw, A. Alfons, P.J.F. Groenen and I. Wilms (2025)
\emph{Clusterpath Gaussian Graphical Modeling}. arXiv:2407.00644.
doi:10.48550/arXiv.2407.00644.
}
\seealso{
\code{\link{clusterpath_weights}()}, \code{\link{lasso_weights}()},
\code{\link{cggm}()}, \code{\link{cggm_refit}()}
}
\author{
Daniel J.W. Touw, modifications by Andreas Alfons
}
