% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cggm_refit.R
\name{cggm_refit}
\alias{cggm_refit}
\title{Refit the Gaussian Graphical Model for a Given Aggregation and Sparsity Structure}
\usage{
cggm_refit(cggm_output, verbose = 0)
}
\arguments{
\item{cggm_output}{Output of \code{\link{cggm}()}.}

\item{verbose}{Determines the amount of information printed during the
optimization. Defaults to \code{0}.}
}
\value{
A list containing the estimated parameters of the clustered,
unpenalized CGGM model.
}
\description{
Estimate the parameters of a clustered and sparse precision matrix or
covariance matrix based on a restricted negative log-likelihood loss
function. The restrictions are given by the provided aggregation and
sparsity structure. This function is different from \code{\link{cggm}()},
as there are no aggregation and sparsity penalties on the precision or
covariance matrix.
}
\references{
D.J.W. Touw, A. Alfons, P.J.F. Groenen and I. Wilms (2025)
\emph{Clusterpath Gaussian Graphical Modeling}. arXiv:2407.00644.
doi:10.48550/arXiv.2407.00644.
}
\seealso{
\code{\link{cggm}()}, \code{\link{cggm_cv}()}
}
\author{
Daniel J.W. Touw
}
