% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cggm_refit.R
\name{cggm_refit}
\alias{cggm_refit}
\title{Estimate Clustered Gaussian Graphical Model without Penalty}
\usage{
cggm_refit(cggm_output, verbose = 0)
}
\arguments{
\item{cggm_output}{Output of \code{\link{cggm}}.}

\item{verbose}{Determines the amount of information printed during the
optimization. Defaults to \code{0}.}
}
\value{
A list containing the estimated parameters of the clustered,
unpenalized CGGM model.
}
\description{
This function estimates the parameters of a clustered precision matrix based
on a restricted negative log-likelihood loss function. The restriction is the
clustering provided by the input. This function is different from
\code{\link{cggm}}, as there is no penalization of the differences between
the different rows of Theta.
}
\references{
D.J.W. Touw, A. Alfons, P.J.F. Groenen and I. Wilms (2025)
\emph{Clusterpath Gaussian Graphical Modeling}. arXiv:2407.00644.
doi:10.48550/arXiv.2407.00644.
}
\seealso{
\code{\link{cggm}}, \code{\link{cggm_cv}}
}
\author{
Daniel J.W. Touw
}
