% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cggmRefit.R
\name{cggmRefit}
\alias{cggmRefit}
\title{Estimate Clusterpath Gaussian Graphical Model}
\usage{
cggmRefit(
  cggm_output,
  S,
  gss_tol = 1e-04,
  conv_tol = 1e-09,
  max_iter = 2000,
  use_Newton = TRUE,
  verbose = 0
)
}
\arguments{
\item{cggm_output}{Output of cggm().}

\item{S}{The sample covariance matrix of the data.}

\item{gss_tol}{The tolerance value used in the Golden Section Search (GSS)
algorithm. Defaults to \code{1e-4}.}

\item{conv_tol}{The tolerance used to determine convergence. Defaults to
\code{1e-9}.}

\item{max_iter}{The maximum number of iterations allowed for the optimization
algorithm. Defaults to \code{2000}.}

\item{use_Newton}{Logical, indicating whether to use Newton's method in the
optimization algorithm. Defaults to \code{TRUE}.}

\item{verbose}{Determines the amount of information printed during the
optimization. Slows down the algorithm significantly. Defaults to \code{0}.}
}
\value{
A list containing the estimated parameters of the CGGM model.
}
\description{
This function estimates the parameters of a CGGM model based on the input
data and parameters.
}
\examples{
# Example usage:

}
