% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterglasso.R
\name{clusterglasso}
\alias{clusterglasso}
\title{tag-lasso estimation of the precision matrix}
\usage{
clusterglasso(
  X,
  pendiag = F,
  lambda1,
  lambda2,
  rho = 10^-2,
  it_in = 500,
  it_out = 10
)
}
\arguments{
\item{X}{An (\eqn{n}x\eqn{p})-matrix of \eqn{p} variables and \eqn{n} observations}

\item{pendiag}{Logical indicator whether or not to penalize the diagonal in Omega. The default is \code{TRUE} (penalization of the diagonal)}

\item{lambda1}{Sparsity tuning parameter.}

\item{lambda2}{Aggregation tuning parameter.}

\item{rho}{Starting value for the LA-ADMM tuning parameter. Default is 10^2; will be locally adjusted via LA-ADMM}

\item{it_in}{Number of inner stages of the LA-ADMM algorithm. Default is 500}

\item{it_out}{Number of outer stages of the LA-ADMM algorithm. Default is 10}
}
\value{
A list with the following components
\item{\code{omega_full}}{Estimated (\eqn{p}x\eqn{p}) precision matrix}
\item{\code{cluster}}{Numeric vector indicating the cluster groups for each of the \eqn{p} original variables}
\item{\code{sparsity}}{The (\eqn{p}x\eqn{p} matrix indicating the sparsity pattern in Omega (1=non-zero, 0=zero))}
\item{\code{fit}}{Fitted object from LA_ADMM_clusterglasso_export cpp function, for internal use now}
}
\description{
This function computes the tag-lasso estimator for fixed tuning parameters lambda1 and lambda2
}
