% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_clusters.R
\name{min_clusters}
\alias{min_clusters}
\title{Calculate the Minimum Number of Clusters}
\usage{
min_clusters(W)
}
\arguments{
\item{W}{The weight matrix used in the clusterpath penalty.}
}
\value{
An integer representing the minimum number of clusters.
}
\description{
Compute the minimum number of clusters achievable using the provided weight
matrix.
}
