% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taglasso.R
\name{taglasso}
\alias{taglasso}
\title{Tree-Aggregated Graphical Lasso}
\usage{
taglasso(
  X,
  A,
  pendiag = F,
  lambda1,
  lambda2,
  adaptive = FALSE,
  power_adaptive = 1,
  W_sparsity = NULL,
  rho = 10^-2,
  it_in = 100,
  it_out = 10,
  refitting = T,
  it_in_refit = 100,
  it_out_refit = 10
)
}
\arguments{
\item{X}{An (\eqn{n}x\eqn{p})-matrix of \eqn{p} variables and \eqn{n} observations}

\item{A}{An (\eqn{p}x\eqn{|T|})- binary matrix incorporating the tree-based aggregation structure}

\item{pendiag}{Logical indicator whether or not to penalize the diagonal in Omega. The default is \code{FALSE} (no penalization of the diagonal)}

\item{lambda1}{Aggregation tuning parameter. Use \code{taglasso_cv} to let the program determine this tuning parameter based on K-fold cross-validation}

\item{lambda2}{Sparsity tuning parameter. Use \code{taglasso_cv} to let the program determine this tuning parameter based on K-fold cross-validation}

\item{adaptive}{Logical indicator whether an adaptive lasso type of sparsity penalty should be used or not. Default is FALSE}

\item{power_adaptive}{Power for the weights int he adaptive lasso type of sparsity penalty. Default is one.}

\item{W_sparsity}{An (\eqn{p}x\eqn{p})-matrix of weights for the sparsity penalty term. Only relevant if adaptive = TRUE}

\item{rho}{Starting value for the LA-ADMM tuning parameter. Default is 10^2; will be locally adjusted via LA-ADMM}

\item{it_in}{Number of inner stages of the LA-ADMM algorithm. Default is 100}

\item{it_out}{Number of outer stages of the LA-ADMM algorithm. Default is 10}

\item{refitting}{Logical indicator whether refitting subject to sparsity and aggregation constraints is done or not. The default is \code{TRUE}.}

\item{it_in_refit}{Number of inner stages of the LA-ADMM algorithm for re-fitting (Only relevant if \code{refitting = TRUE}). Default is 100}

\item{it_out_refit}{Number of outer stages of the LA-ADMM algorithm for re-fitting (Only relevant if \code{refitting = TRUE}). Default is 10}
}
\value{
A list with the following components:
\item{omega_full}{Estimated (\eqn{p}x\eqn{p}) precision matrix}
\item{cluster}{Numeric vector indicating the cluster groups for each of the \eqn{p} original variables}
\item{sparsity}{The (\eqn{p}x\eqn{p} matrix indicating the sparsity pattern in Omega (1=non-zero, 0=zero))}
\item{fit}{Fitted object from LA_ADMM_taglasso_export cpp function, for internal use now}
\item{refit}{Fitted object from refit_LA_ADMM_export cpp function, for internal use now}
}
\description{
Compute the tree-aggregated graphical lasso (tag-lasso) estimator of the precision matrix for fixed values of the tuning parameters.
}
\note{
The function interface and output structure are still experimental and
may change in the next version.
}
\references{
I. Wilms and J. Bien (2022) Tree-based Node Aggregation in Sparse Graphical
Models. \emph{Journal of Machine Learning Research}, \bold{23}(243), 1--36.
https://jmlr.org/papers/v23/21-0105.html
}
\author{
Ines Wilms and Jacob Bien
}
